@@ -269,13 +269,7 @@ namespace System.Management.Automation
         /// </param>
         internal PSTraceSource(string fullName, string name, string description, bool traceHeaders)
         {
-            if (string.IsNullOrEmpty(fullName))
-            {
-                // 2005/04/13-JonN In theory this should be ArgumentException,
-                // but I don't want to deal with loading the string in this
-                // low-level code.
-                throw new ArgumentNullException(nameof(fullName));
-            }
+            ArgumentException.ThrowIfNullOrEmpty(fullName);
 
             try
             {
