@@ -523,7 +523,7 @@ namespace System.Management.Automation
         internal static string ComputeHash(string input)
         {
             byte[] hashBytes = ComputeHash(Encoding.UTF8.GetBytes(input));
-            return BitConverter.ToString(hashBytes).Replace("-", string.Empty);
+            return Convert.ToHexString(hashBytes);
         }
     }
 
